+++ src/main/java/pulleydoreurae/careerquestbackend/auth/domain/dto/request/UserAccountRegisterRequest.java
@@ -1,4 +1,4 @@
-package pulleydoreurae.careerquestbackend.auth.domain.dto;
+package pulleydoreurae.careerquestbackend.auth.domain.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
