+++ src/main/java/pulleydoreurae/careerquestbackend/auth/domain/dto/response/KakaoUserDetailsResponse.java
@@ -1,10 +1,11 @@
-package pulleydoreurae.careerquestbackend.auth.domain.dto;
+package pulleydoreurae.careerquestbackend.auth.domain.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import pulleydoreurae.careerquestbackend.auth.domain.dto.request.KakaoAccountRequest;
 
 /**
  * 카카오 로그인에 성공한 토큰으로 사용자 정보를 받아올 dto
@@ -21,5 +22,5 @@ import lombok.Setter;
 public class KakaoUserDetailsResponse {
 
 	private Long id;
-	private KakaoAccount kakao_account;
+	private KakaoAccountRequest kakao_account;
 }
