+++ src/main/java/pulleydoreurae/careerquestbackend/auth/domain/dto/response/UserAccountRegisterResponse.java
@@ -1,4 +1,4 @@
-package pulleydoreurae.careerquestbackend.auth.domain.dto;
+package pulleydoreurae.careerquestbackend.auth.domain.dto.response;
 
 import lombok.Builder;
 import lombok.Getter;
